email-security-system/backend/requirements.txt
Flask
SQLAlchemy
Flask-JWT-Extended
Flask-Mail
Flask-CORS
Flask-SocketIO
requests
psycopg2-binary
python-dotenv
scikit-learn
numpy
eventlet 